List groups with group_type_id 2 (= "PWG") as "BOF".
This makes conex and ppsp come out correctly as BOFs (instead of nameless) in
the IETF77 agenda.

But maybe these should be indicated differently?
It is not clear what a PWG is meant to be; maybe WGs-to-be scheduled
as BOFs?
Once this is clarified, this changed needs to be reviewed.
 - Legacy-Id: 2148