Use the full_draft cookie to allow users to store their preference to see the whole draft on the document page.
Patch django to use the time.time hack that Mark Nottingham sent
me, to avoid putting the current time in the gzip stream, which
allows the content to not change, enabling caching.

Add Cache-Control: private, max-age=3600 headers to the
document pages, to encourage caching.  Tell django to generate
ETag headers and handle If-None-Match or Not-Modified-Since
internally.
 - Legacy-Id: 2685